MonsterCrud - added two examples of address_algolia instead of one
Since there are two cases (stored as JSON and stored as STRING), we now have examples in the MonsterCrud for both use cases. Made the column text instead of string because it caused a massive problem sometimes that was difficult to debug. When the JSON was too large, it got trimmed by the database, so when it was stored it was an invalid JSON. So when the field tried to json_decode it, it wasn't valid, it looked like NULL. This made the field stop working in Edit/Update, without an apparent reason. Seemed like the field was broken, but no - in fact the db column could not hold the entire string. Now fixed in the demo. WIll mention in the docs too, so that others don't get the same frustrating DX.